rename parseData.js to ParseNode.js and export ParseNode as the default
